requirements.txt
flask==3.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# 報表生成與繪圖
weasyprint==61.2
matplotlib==3.8.2
pandas==2.1.3
numpy==1.26.2

# 前端 (Streamlit)
streamlit==1.29.0
requests==2.31.0

# 郵件寄送
yagmail==0.15.293

# 其他常用工具
python-dotenv==1.0.0